create journey options component
