+++ src/routes/journey/+page.svelte
@@ -5,19 +5,11 @@
 		type DisplayedFormData,
 		displayedFormData,
 		initializeSharedData,
-		refreshJourney,
 		selectedJourneys
 	} from "$lib/stores/journeyStores";
-	import IconRefresh from "$lib/components/icons/IconRefresh.svelte";
 	import { page } from "$app/stores";
 	import { browser } from "$app/environment";
-	import { shareJourney } from "./share";
-	import IconShare from "$lib/components/icons/IconShare.svelte";
-	import { type ComponentProps, onMount, type Snippet } from "svelte";
-	import { getBookmarks, type JourneyBookmark, toggleJourneyBookmark } from "$lib/bookmarks";
-	import IconBookmark from "$lib/components/icons/IconBookmark.svelte";
-	import { getJourneyUrl } from "$lib/urls";
-	import Options from "$lib/components/Options.svelte";
+	import { type Snippet } from "svelte";
 	import IconJourneyInfo from "$lib/components/icons/IconJourneyInfo.svelte";
 	import IconMap from "$lib/components/icons/IconMap.svelte";
 	import MiniTabs from "$lib/components/MiniTabs.svelte";
@@ -26,7 +18,7 @@
 	import { dateToString, timeToString } from "$lib/util.js";
 	import { settings } from "$lib/stores/settingStore";
 	import TicketModal from "$lib/components/TicketModal.svelte";
-	import IconTickets from "$lib/components/icons/IconTickets.svelte";
+	import JourneyOptions from "./JourneyOptions.svelte";
 
 	const { formData, treeNodes } = $page.data;
 
@@ -55,62 +47,7 @@
 	let clientWidth: number = $state(0);
 
 	let allSelected = $derived($selectedJourneys.every((journey) => journey.selectedBy !== -1));
-
-	let journeyBookmarks: JourneyBookmark[] = $state([]);
-
-	onMount(() => (journeyBookmarks = getBookmarks("journey")));
-
-	let isBookmarked = $derived(
-		browser &&
-			journeyBookmarks.some(
-				(bookmark) => bookmark.link === getJourneyUrl($selectedJourneys).href
-			)
-	);
-
-	function handleJourneyBookmarkClick(): void {
-		journeyBookmarks = toggleJourneyBookmark($selectedJourneys, $displayedFormData);
-	}
-
-	const options: ComponentProps<typeof Options>["options"] = [
-		{
-			type: "function",
-			name: "Aktualisieren",
-			icon: iconRefresh,
-			onClick: refreshJourney
-		},
-		{
-			type: "function",
-			name: "Teilen",
-			icon: iconShare,
-			onClick: () => shareJourney($selectedJourneys)
-		},
-		{
-			type: "function",
-			name: "Merken",
-			icon: iconBookmark,
-			onClick: handleJourneyBookmarkClick
-		},
-		{
-			type: "modal",
-			name: "Tickets",
-			icon: iconTickets,
-			showModalKey: "showTicketModal"
-		}
-	];
 </script>
-
-{#snippet iconRefresh()}
-	<IconRefresh />
-{/snippet}
-{#snippet iconBookmark()}
-	<IconBookmark {isBookmarked} />
-{/snippet}
-{#snippet iconShare()}
-	<IconShare />
-{/snippet}
-{#snippet iconTickets()}
-	<IconTickets />
-{/snippet}
 
 <svelte:head>
 	<title>Vahrplan - Reisedetails{pageTitle.length > 0 ? ": " : ""}{pageTitle}</title>
@@ -160,9 +97,7 @@
 		{#snippet tabEnvironment(miniTabSelector: Snippet, tabContent: Snippet)}
 			<Header title={pageTitle.length === 0 ? "Reisedetails" : pageTitle} mobileOnly={true}>
 				{@render miniTabSelector()}
-				{#if allSelected && $selectedJourneys.length > 0}
-					<Options id={"journey"} {options} />
-				{/if}
+				<JourneyOptions />
 			</Header>
 			<div class="content-wrapper padded-top-bottom">
 				{@render tabContent()}
@@ -173,26 +108,9 @@
 	<div class="columns">
 		<section class="journeys content-wrapper">
 			<TitlelessHeader>
-				{#if allSelected && $selectedJourneys.length > 0}
-					<div class="flex-row journey-actions--buttons">
-						<TicketModal />
-						<button
-							class="hoverable hoverable--visible"
-							onclick={() => void shareJourney($selectedJourneys)}
-						>
-							<IconShare />
-						</button>
-						<button
-							class="hoverable hoverable--visible"
-							onclick={handleJourneyBookmarkClick}
-						>
-							<IconBookmark {isBookmarked} />
-						</button>
-						<button class="hoverable hoverable--visible" onclick={refreshJourney}>
-							<IconRefresh />
-						</button>
-					</div>
-				{/if}
+				<div class="journey-options">
+					<JourneyOptions />
+				</div>
 			</TitlelessHeader>
 			{@render journeyOverview()}
 		</section>
@@ -220,9 +138,7 @@
 		max-height: 100%;
 		overflow: auto;
 	}
-	.journey-actions--buttons {
+	.journey-options {
 		padding: var(--line-width) 0.75rem 0;
-		gap: var(--line-width);
-		justify-content: end;
 	}
 </style>
